Split groovydoc plugin into generic base and Grails layer
Separate GrailsGroovydocPlugin into a generic GroovydocEnhancerPlugin
(with GroovydocEnhancerExtension) and a thin GrailsGroovydocPlugin that
applies the base plugin and sets the Matomo footer. This makes the core
AntBuilder groovydoc logic publishable and reusable by anyone, while
keeping Grails-specific customizations in their own layer.

The base plugin supports a useAntBuilder flag so projects can easily
switch back to Gradle's built-in Groovydoc task if Gradle merges their
javaVersion support (gradle/gradle#33659).

Also fix a pre-existing bug in resolveProjectVersion() in
docs-dependencies.gradle where the function never returned the resolved
version string, causing all external groovydoc links (geb,
testcontainers, spring, spring-boot) to be silently omitted.
